changes to build system & fixing up warnings

+++ source/FlxColor.cpp
@@ -8,9 +8,9 @@ Flx::Color Flx::Color::fromHex(int color)
 {
     return Color
     {
-        ((color >> 24) & 0xFF), 
-        ((color >> 16) & 0xFF), 
-        ((color >> 8) & 0xFF), 
-        ((color) & 0xFF)
+        static_cast<uint8_t>((color >> 24) & 0xFF), 
+        static_cast<uint8_t>((color >> 16) & 0xFF), 
+        static_cast<uint8_t>((color >> 8) & 0xFF), 
+        static_cast<uint8_t>((color) & 0xFF)
     };
 }
